Major refactoring is done.
The window still starts up as a postage stamp when it first starts up -- that popped up with my changes. I've been trying to nail down issues with the new CursesLikeAPI class first. Since that now seems to be in order, the next step is to fix the postage stamp issue and work on performance.

+++ core/src/test/java/com/googlecode/blacken/colors/TestColorHelper.java
@@ -90,8 +90,10 @@ public class TestColorHelper {
      */
     @Test
     public void makeVisible_Value_Threshold_White_Black() {
-        assertEquals(RED, ColorHelper.makeVisible(GREY_HALF, 75, RED, BLUE));
-        assertEquals(BLUE, ColorHelper.makeVisible(GREY_HALF, 25, RED, BLUE));
+        assertEquals(RED, 
+                (int)ColorHelper.makeVisible(GREY_HALF, 75, RED, BLUE));
+        assertEquals(BLUE, 
+                (int)ColorHelper.makeVisible(GREY_HALF, 25, RED, BLUE));
     }
     
     /**
